% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/density.R
\docType{data}
\name{density}
\alias{density}
\title{Density metrics}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 222 rows and 4726 columns.
}
\usage{
density
}
\description{
density (ω_j^c ) of country c in terms of each product from CEPPII BACI trade data
}
\references{
(https://link.springer.com/article/10.1057/s42214-020-00081-7)
}
\author{
R Theron \email{rolandingenieur98@gmail.com}
}
\keyword{data}
