% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\docType{data}
\name{distance}
\alias{distance}
\title{Distance metrics}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 222 rows and 4726 columns.
}
\usage{
distance
}
\description{
distance (∆_j^c) of a product j in which country c does not yet have an RCA. From CEPPII BACI trade data
}
\references{
(https://link.springer.com/article/10.1057/s42214-020-00081-7)
}
\author{
R Theron \email{rolandingenieur98@gmail.com}
}
\keyword{data}
