% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ECI.R
\docType{data}
\name{ECI}
\alias{ECI}
\title{Economic Complexity Indices}
\format{
An object of class \code{numeric} of length 222.
}
\usage{
ECI
}
\description{
Economic complexity indices for every ISO 3-digit country code from CEPII BACI trade data
}
\references{
(https://link.springer.com/article/10.1057/s42214-020-00081-7)
}
\author{
R Theron \email{rolandingenieur98@gmail.com}
}
\keyword{data}
