% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExampleTradeData.R
\docType{data}
\name{ExampleTradeData}
\alias{ExampleTradeData}
\title{BACI_HS17_Y2020_V202301 data}
\format{
`ExampleTradeData`
A data frame with 72097 rows and 6 variables:
\describe{
  \item{t}{Year}
  \item{k}{Product category, HS 6-digit code}
  \item{i}{Exporter, ISO 3-digit country code}
  \item{j}{Importer, ISO 3-digit country code}
  \item{v}{Value of the trade flow, in thousands current USD}
  \item{q}{Quantity, in metric tons}
}
}
\source{
\url{http://www.cepii.fr/CEPII/en/bdd_modele/bdd_modele_item.asp?id=37}
}
\usage{
ExampleTradeData
}
\description{
A dataset containing the trade data for country codes 4 and 710.
}
\keyword{datasets}
