% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/M_absolute.R
\docType{data}
\name{M_absolute}
\alias{M_absolute}
\title{Absolute M-matrix}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 222 rows and 4726 columns.
}
\usage{
M_absolute
}
\description{
Absolute M-matrix of the CEPPII BACI trade data
}
\references{
(https://link.springer.com/article/10.1057/s42214-020-00081-7)
}
\author{
R Theron \email{rolandingenieur98@gmail.com}
}
\keyword{data}
