% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Opportunity_Gain.R
\docType{data}
\name{Opportunity_Gain}
\alias{Opportunity_Gain}
\title{Opportunity Gain}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 222 rows and 4726 columns.
}
\usage{
Opportunity_Gain
}
\description{
Opportunity gain for opportunity j can be defined as the change in country c's opportunity value if an RCA from for opportunity j is obtained.
}
\references{
(https://link.springer.com/article/10.1057/s42214-020-00081-7)
}
\author{
R Theron \email{rolandingenieur98@gmail.com}
}
\keyword{data}
