% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCI.R
\docType{data}
\name{PCI}
\alias{PCI}
\title{Product Complexity Indices}
\format{
An object of class \code{numeric} of length 4726.
}
\usage{
PCI
}
\description{
Product complexity indices for every HS product code from CEPII BACI trade data
}
\references{
(https://link.springer.com/article/10.1057/s42214-020-00081-7)
}
\author{
R Theron \email{rolandingenieur98@gmail.com}
}
\keyword{data}
