% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOPS_Full_ALPHA.R
\name{IOPS}
\alias{IOPS}
\title{IOPS}
\usage{
IOPS(
  CountryCode,
  tradeData,
  ComplexMethod = "eigenvalues",
  iterCompl = 20,
  GVCMapping
)
}
\arguments{
\item{CountryCode}{(Type: character/integer) Any accepted ISO country code could be used, e.g. \code{"United Kingdom"}, \code{"GBR"}, \code{"GB"}, \code{828} would all be accepted if the United Kingdom is the desired country.}

\item{tradeData}{(Type: csv) Accepts any CEPII BACI trade data. NOTE: tradeData and GVCMapping must be from the same "H" Family, e.g. both are from  H3, etc., in order for the program to work correctly.}

\item{ComplexMethod}{(Type: character) Methods used to calculate complexity measures. Can be any one of these methods: \code{"fitness"}, \code{"reflections"} or \code{"eigenvalues"}. Defaults to "eigenvalues".}

\item{iterCompl}{(Type: integer) The number of iterations that the chosen complexity measure must use. Defaults to \code{iterCompl = 20}.}

\item{GVCMapping}{(Type: csv) The desired value chain to be analysed. With Columns "Tiers", "Activity", and "HSCode". NOTE: tradeData and GVCMapping must be from the same "H" Family, e.g. both are from  H3, etc., in order for the program to work correctly.}
}
\value{
A list that constains ECI, PCI, Opportunity_Gain, distance, density, M_absolute, M_binary, Tier_Results, Product_Category_Results, Product_Results, respectively.
}
\description{
Takes user inputted trade data, any acceptable ISO country code and industrial value chain mapping to calculate various metrics (Economic- and Product complexity metrics, distance metrics, opportunity gain, and inequality metrics) of a given country in order to facilitate better decision making regarding industrial policymaking.
}
